requirements.txt
discord.py
asyncio
PyNaCl
ffmpeg
ffmpeg-python
requests
youtube-dl